* Make 'rasm' happy with -a rsc   - Make 'rsc asm' use SYNTAX=intel by default   - add 'olly' as available arch in help   - fix performance bug parsing arches * Import waf-svn script at root   - Fix some structural build problems   - Make it use global.h   - Fix python plugin build on ubuntu   - Fix some warnings   - add --with-maemo and --with-sysproxy   $ ./waf configure --without-debugger   $ ./waf

+++ src/plug/io/malloc.c
@@ -31,7 +31,7 @@ static unsigned int malloc_bufsz = 0;
 
 static ssize_t malloc_write(int fd, const void *buf, size_t count)
 {
-	return memcpy(malloc_buf+config.seek, buf, count);
+	return (ssize_t)memcpy(malloc_buf+config.seek, buf, count);
 }
 
 static ssize_t malloc_read(int fd, void *buf, size_t count)
@@ -44,7 +44,7 @@ static ssize_t malloc_read(int fd, void *buf, size_t count)
 	if (config.seek +count > config.size)
 		config.seek = config.size;
 
-	return memcpy(buf, malloc_buf+config.seek, count);
+	return (ssize_t)memcpy(buf, malloc_buf+config.seek, count);
 }
 
 static int malloc_close(int fd)
